Actualizada linea conflictiva del tokenizer en pwd cd y exit. y ejemplos de comandos en el .txt

+++ MiniShell/tokenizer.c
@@ -117,12 +117,12 @@ void tokens_destroy(struct tokens *tokens) {
   if (tokens == NULL) {
     return;
   }
-  for (int i = 0; i < tokens->tokens_length; i++) {
+  for (size_t i = 0; i < tokens->tokens_length; i++) {
     free(tokens->tokens[i]);
   }
   free(tokens->tokens);
 
-  for (int i = 0; i < tokens->buffers_length; i++) {
+  for (size_t i = 0; i < tokens->buffers_length; i++) {
     free(tokens->buffers[i]);
   }
   free(tokens->buffers);
